+++ inst/doc/tips-tricks.R
@@ -40,42 +40,16 @@ create_notebook_explore(
 #               output_dir = tempdir())
 
 ## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- use_data_titanic(count = FALSE)
-glimpse(data)
+colors <- mix_color("blue", n = 5)
+colors
 
 ## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- data %>% clean_var(Age, name = "age")
-glimpse(data)
+show_color(colors)
 
 ## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- use_data_beer()
-data %>% describe(energy_kcal_100ml)
+colors <- mix_color("gold", "red", n = 4)
+colors
 
 ## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- data %>% clean_var(energy_kcal_100ml, na = 42)
-data %>% describe(energy_kcal_100ml)
-
-## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- create_data_person()
-data %>% describe(age)
-
-## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- data %>% clean_var(age, min_val = 20, max_val = 80)
-data %>% describe(age)
-
-## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data %>% describe(income)
-
-## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- data %>% clean_var(income, rescale01 = TRUE)
-data %>% describe(income)
-
-## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data[1, "handset"] <- " android "
-data[2, "handset"] <- "ANDROID"
-data %>% describe(handset)
-
-## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
-data <- data %>% clean_var(handset, simplify_text = TRUE)
-data %>% describe(handset)
+show_color(colors)